app/src/test/java/be/ugent/zeus/hydra/common/utils/PreferencesUtilsTest.java
/*
 * Copyright (c) 2021 The Hydra authors
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package be.ugent.zeus.hydra.common.utils;

import android.content.Context;
import android.content.SharedPreferences;
import androidx.preference.PreferenceManager;
import androidx.test.core.app.ApplicationProvider;

import java.util.Arrays;
import java.util.HashSet;
import java.util.Set;

import be.ugent.zeus.hydra.TestApp;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.annotation.Config;

import static org.hamcrest.CoreMatchers.not;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.contains;
import static org.hamcrest.Matchers.containsInAnyOrder;

/**
 * @author Niko Strijbol
 */
@RunWith(RobolectricTestRunner.class)
@Config(application = TestApp.class)
public class PreferencesUtilsTest {

    private static final String TEST_KEY = "test_key";

    private Context context;

    @Before
    public void setUp() {
        this.context = ApplicationProvider.getApplicationContext();
    }

    @Test
    public void addToStringSet() {
        String value = "1";
        PreferencesUtils.addToStringSet(context, TEST_KEY, value);
        Set<String> result = PreferencesUtils.getStringSet(context, TEST_KEY);
        assertThat(result, contains(value));
    }

    @Test
    public void removeFromStringSet() {
        String[] data = {"1", "2", "3", "4", "a", "5"};

        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(context);
        preferences.edit()
                .putStringSet(TEST_KEY, new HashSet<>(Arrays.asList(data)))
                .apply();

        PreferencesUtils.removeFromStringSet(context, TEST_KEY, "a");
        Set<String> result = PreferencesUtils.getStringSet(context, TEST_KEY);
        assertThat(result, not(contains("a")));
        assertThat(result, containsInAnyOrder("1", "2", "3", "4", "5"));
    }
}
